[EI-260] [web] refactor: Update imports to use the new use-logger.hook
