+++ apps/web/app/business/hooks/linguistic-guidance/use-analyze-economy.hook.ts
@@ -1,4 +1,4 @@
-import { useLogger } from '@/app/logging/logging-context';
+import { useLogger } from '@/app/logging/use-logger.hook';
 import { addIndicatorsToMetadata, getIndicatorValue } from '../../services/linguistic-guidance/indicator.service';
 import {
   getIndicatorBySymbol,
