+++ apps/web/app/ui/pages/workspace/split-screen/split-screen-toggle-group.tsx
@@ -2,7 +2,7 @@
 import { useIndicatorBoard } from '@/app/business/hooks/numerical-guidance/indicator-board/use-indicator-board.hook';
 import ToggleGroup from '../../../components/view/molecule/toggle-group';
 import { type SplitScreen, splitScreens } from '@/app/store/stores/numerical-guidance/indicator-board.store';
-import { useLogger } from '@/app/logging/logging-context';
+import { useLogger } from '@/app/logging/use-logger.hook';
 
 function isSplitScreen(value: string): value is SplitScreen {
   return splitScreens.includes(value as SplitScreen);
